Got initial Comet based on long pool working.
We keep a map of cometIds (could be a slice), and we loop over them and send new
chat messages to them. This also takes care of comets that are not on the
page any more, as we remove them from the map as soon as
we send them a message.
The comets that are still listening, will trigger a new long pool and be added to the map again.

Once go-restful implements the CloseNotify() interface, we could clean this up